+++ core/security/src/iosMain/kotlin/com/chimali/core/security/platform/ResourceLoader.ios.kt
@@ -1,0 +1,5 @@
+package com.chimali.core.security.platform
+
+actual fun loadResourceLines(name: String): List<String> {
+    TODO("iOS resource loading not yet implemented")
+}